Продолжительность фильма указан как double вместо duration

+++ src/main/java/ru/yandex/practicum/filmorate/validate/Validation.java
@@ -91,8 +91,8 @@ public class Validation {
         }
     }
 
-    public boolean positiveDurationFilm(int duration) {
-        if (duration > 0) {
+    public boolean positiveDurationFilm(double duration) {
+        if (duration > 0.0) {
             return true;
         } else {
             log.info("Продолжительность фильма должна быть положительным числом");
